test(runtime): repoint rollback cleanup doubles at schedule_cleanup
The worker finalization path now calls RunManager.schedule_cleanup, so the
store-less CleanupTrackingRunManager doubles in test_run_worker_rollback.py
that only overrode cleanup() no longer observe any call:

- test_run_agent_schedules_terminal_cleanup_when_publish_end_fails now
  tracks schedule_cleanup directly (keeps its bridge and GC assertions).
- The completion-hook-cancelled and MCP-projection-cancelled scenarios move
  to test_run_worker_finalization.py alongside the other eviction pins.
- test_run_agent_schedules_terminal_run_record_cleanup is dropped as a
  duplicate of test_run_agent_finalization_schedules_eviction.
- test_terminal_cleanup_tasks_do_not_inherit_run_context is dropped;
  test_run_manager.py already pins the empty-Context behavior.
